Separate EncryptionTokenProvider from DeviceTokenProvider
