--- auth-foundation/src/androidTest/java/com/okta/authfoundation/credential/LegacyDeviceTokenProviderTest.kt
@@ -1,57 +1,0 @@
-/*
- * Copyright 2023-Present Okta, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.okta.authfoundation.credential
-
-import android.content.Context
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.common.truth.Truth.assertThat
-import com.okta.authfoundation.client.LegacyDeviceTokenProvider
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-class LegacyDeviceTokenProviderTest {
-    private lateinit var context: Context
-    private lateinit var legacyDeviceTokenProvider: LegacyDeviceTokenProvider
-
-    @Before
-    fun setup() {
-        context = ApplicationProvider.getApplicationContext()
-        legacyDeviceTokenProvider = LegacyDeviceTokenProvider(context)
-        clearSharedPreferences()
-    }
-
-    private fun clearSharedPreferences() {
-        context.getSharedPreferences(LegacyDeviceTokenProvider.FILE_NAME, Context.MODE_PRIVATE).edit().clear().commit()
-    }
-
-    @Test
-    fun testDeviceTokenProviderInitializesCorrectly() {
-        val actualDeviceToken = legacyDeviceTokenProvider.deviceToken
-        val expectedDeviceToken = legacyDeviceTokenProvider.sharedPrefs.getString(
-            LegacyDeviceTokenProvider.PREFERENCE_KEY, null
-        )!!.filter { it.isLetterOrDigit() }
-        assertThat(actualDeviceToken).isEqualTo(expectedDeviceToken)
-    }
-
-    @Test
-    fun testDeviceTokenIsAtMost32Characters() {
-        val actualDeviceToken = legacyDeviceTokenProvider.deviceToken
-        assertThat(actualDeviceToken.length).isAtMost(32)
-    }
-}
