--- auth-foundation/src/test/java/com/okta/authfoundation/client/DeviceTokenProviderTest.kt
@@ -1,81 +1,0 @@
-/*
- * Copyright 2024-Present Okta, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.okta.authfoundation.client
-
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.common.truth.Truth.assertThat
-import com.okta.testhelpers.MockAesEncryptionHandler
-import io.mockk.every
-import io.mockk.mockkConstructor
-import io.mockk.mockkObject
-import io.mockk.unmockkAll
-import kotlinx.coroutines.test.runTest
-import org.junit.After
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-class DeviceTokenProviderTest {
-    private lateinit var deviceTokenProvider: DeviceTokenProvider
-
-    @Before
-    fun setUp() {
-        mockkObject(ApplicationContextHolder)
-        every { ApplicationContextHolder.appContext } returns ApplicationProvider.getApplicationContext()
-
-        mockkConstructor(LegacyDeviceTokenProvider::class)
-        every { anyConstructed<LegacyDeviceTokenProvider>().containsDeviceToken() } returns false
-        deviceTokenProvider = DeviceTokenProvider(MockAesEncryptionHandler.instance)
-    }
-
-    @After
-    fun tearDown() {
-        unmockkAll()
-    }
-
-    @Test
-    fun `getDeviceToken success`() = runTest {
-        val deviceToken = deviceTokenProvider.getDeviceToken()
-        assertThat(deviceToken).isNotEmpty()
-    }
-
-    @Test
-    fun `getDeviceToken is at most 32 characters`() = runTest {
-        val deviceToken = deviceTokenProvider.getDeviceToken()
-        assertThat(deviceToken.length).isAtMost(32)
-    }
-
-    @Test
-    fun `getDeviceToken imports device token from LegacyDeviceTokenProvider`() = runTest {
-        every { anyConstructed<LegacyDeviceTokenProvider>().containsDeviceToken() } returns true
-        every { anyConstructed<LegacyDeviceTokenProvider>().deviceToken } returns "legacyDeviceToken"
-
-        val deviceToken = deviceTokenProvider.getDeviceToken()
-        assertThat(deviceToken).isEqualTo("legacyDeviceToken")
-    }
-
-    @Test
-    fun `getDeviceToken does not replace existing device token with one from LegacyDeviceTokenProvider`() = runTest {
-        val deviceToken = deviceTokenProvider.getDeviceToken()
-
-        every { anyConstructed<LegacyDeviceTokenProvider>().containsDeviceToken() } returns true
-        every { anyConstructed<LegacyDeviceTokenProvider>().deviceToken } returns "legacyDeviceToken"
-
-        assertThat(deviceToken).isNotEqualTo("legacyDeviceToken")
-    }
-}
